+++ .github/CHANGELOG.md
@@ -63,6 +63,7 @@
 1. [A380X/RMP] Fixed a bug where the PILOT_TRANSMITTER_SET key event would toggle transmission rather than only setting it on - @tracernz (Mike)
 1. [MISC] Refactor handling of unknown enum values to fix issues with GSX "complete now" feature and potential other wasm crashes - @Saschl
 1. [FMS] Fix a bug where the flight number was not retained when entering a new FROM/TO - @tracernz (Mike)
+1. [A380X/MFD] Show RNP AR approaches as "RNP" instead of RNAV - @BravoMike99 (bruno_pt99)
 
 ## 2024.1.0